refactor: State와 하위객체 패키지 이동
